delegete -> function is again compile time :-/ not like it, but c++ standard is a piece of shit. begin of resource manager utility.h -> remove_reference / forward / move

+++ src/core/hash_map.h
@@ -161,7 +161,7 @@ public:
 	{
 		if (m_bucketSize == 0)
 			Rehash(INITIAL_SIZE);
-		if ((m_size / m_bucketSize) > MAX_FACTOR)
+		if ((m_size / (float)m_bucketSize) > MAX_FACTOR)
 			Rehash(m_bucketSize * ENLARGE_MULTIPLIER);
 
 		unsigned bucketIdx = GetIndex(key);
